==> notebooks/01_data_cleaning.ipynb <==
#### Imports

import pandas as pd

#### Loading Dataset

df = pd.read_csv("../NBA_Dataset.csv")
df

## 0) Define unit of analysis + key

KEY = ["player", "season", "age"] 

## 1) Structural Checking

df.head

df.columns = df.columns.str.strip().str.lower()


df.columns

len(df.columns)

df.index

df.dtypes

df.shape

## 2. Data Cleaning/Preprocessing

#### Handling Missing Values

# Remove Players with No Games
df = df[df["g"] > 0].copy()

df.isna().sum()

# Checking if NA values are a MAR (missing at random) or MNAR (missing not at random)
pairs = [
    ("fg_pct",  "fga_per_g"),
    ("fg3_pct", "fg3a_per_g"),
    ("ft_pct",  "fta_per_g"),
    ("fg2_pct", "fg2a_per_g"),
]

for col, denom in pairs:
    print(f"\nChecking {col} vs {denom}")
    print(
        df.loc[df[col].isna(), denom].describe()
    )

# shooting percentages
col = [
    "fg_pct", "fg3_pct","ft_pct", "fg2_pct"
]
df[col] = df[col].fillna(0)

# advanced stats
adv_cols = [
    "per","ts_pct","efg_pct","usg_pct",
    "orb_pct","drb_pct","trb_pct",
    "ast_pct","stl_pct","blk_pct",
    "tov_pct","ws_per_48", "fg3a_per_fga_pct", "fta_per_fga_pct"
]
df[adv_cols] = df[adv_cols].fillna(0)

df.isna().sum()

df.dropna(inplace=True)

df["age"].isna().sum()

#### Checking for duplicates

df["player"] = df["player"].str.strip().str.lower()

# Check for duplicate player-season-age entries
df.duplicated(subset=KEY).sum()

#Checked for same player same season same age but on a different team
(df.groupby(KEY)["team_id"].nunique() > 1).sum()

#### Validating Games per Season

df.loc[df["g"] <= 82, "g"].describe()

df = df[df["g"] <= 82].copy()

df.shape

# Produce an artifact
df.to_parquet("../data/cleaned.parquet", index=False)
